Cleaned up all entities

+++ Backend/src/main/java/dtos/HarbourDTO.java
@@ -4,30 +4,30 @@ import entities.Harbour;
 
 public class HarbourDTO {
 
-    private int id;
+    private Long id;
     private String address;
     private int capacity;
     private String name;
 
     public HarbourDTO(Harbour h) {
-        this.id = h.getId();
+        this.id = Long.valueOf(h.getId());
         this.address = h.getAddress();
         this.capacity = h.getCapacity();
         this.name = h.getName();
     }
 
-    public HarbourDTO(int id, String address, int capacity, String name) {
+    public HarbourDTO(Long id, String address, int capacity, String name) {
         this.id = id;
         this.address = address;
         this.capacity = capacity;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
